Uploader of HTML
Minimum function for uploading HTML source has been set up.
Just only make a file and write contents of preview screen to the file.
Figures will be writen in HTML direcrly.
It might be problem, however, I don't have to make uploader for the figures when I emproy this system.